use a higher abstraction for merges
