+++ src/eventHandlers/pullRequest/index.ts
@@ -1,8 +1,8 @@
 import { context, getOctokit } from '@actions/github';
 
+import { merge } from '../../common/merge';
 import { findPullRequestLastApprovedReview } from '../../graphql/queries';
 import { ReviewEdges } from '../../types';
-import { mutationSelector } from '../../utilities/graphql';
 import { logInfo, logWarning } from '../../utilities/log';
 
 interface PullRequestInformation {
@@ -86,12 +86,10 @@ export const pullRequestHandle = async (
       )}.`,
     );
 
-    await octokit.graphql(
-      mutationSelector(pullRequestInformation.reviewEdges[0]),
-      {
-        commitHeadline: pullRequest.title as string,
-        pullRequestId: pullRequest.node_id as string,
-      },
-    );
+    await merge(octokit, {
+      commitHeadline: pullRequest.title,
+      pullRequestId: pullRequest.node_id,
+      reviewEdge: pullRequestInformation.reviewEdges[0],
+    });
   }
 };
